refactor : service 패키지 추가 및 Result 클래스 ResultService 로 Rules 클래스 RuleService 로 변경
클래스들의 이름 변경
